requirements.txt
apiview==1.3.25
asgiref==3.5.2
certifi==2022.12.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-cors-headers==3.13.0
django-import-export==3.0.1
django-split-settings==1.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==8.5.3
et-xmlfile==1.1.0
flake8==6.0.0
gunicorn==20.1.0
MarkupPy==1.14
mccabe==0.7.0
odfpy==1.4.1
openpyxl==3.0.10
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
typing_extensions==4.4.0
urllib3==1.26.13
uWSGI==2.0.21
xlrd==2.0.1
xlwt==1.3.0
